ALIO-607 TooltipArrowView 파일 분리
