+++ Sources/ALDesign/Component/Tooltip/ALTooltip.swift
@@ -5,24 +5,7 @@
 //  Created by 이옥민 on 2022/04/07.
 //
 
-import Foundation
 import UIKit
-
-public class TooltipArrowView: UIView {
-    override public func draw(_ rect: CGRect) {
-        super.draw(rect)
-        backgroundColor = .clear
-        print(frame)
-        let path = UIBezierPath()
-        path.move(to: .zero)
-        path.addLine(to: CGPoint(x: frame.width / 2, y: frame.height))
-        path.addLine(to: CGPoint(x: frame.width, y: 0))
-        path.close()
-        
-        UIColor(hexString: "#00BEC7").set()
-        path.fill()
-    }
-}
 
 public class ALTooltip: UIView {
     private let arrowDirection: ALTooltipArrowDirection
